backend/requirements.txt
firebase-admin
requests
python-jose
Flask
flask-cors
